Simplify webpack config
The dev and prod files weren't doing a lot of changes so it is clearer
to have these changes inline.
