--- webpack/webpack.config.prod.js
@@ -1,8 +1,0 @@
-/* global require, module */
-const defaults = require( "./webpack.config.default" ).defaults;
-
-const prodConfig = {
-	devtool: "cheap-module-source-map",
-};
-
-module.exports = defaults( prodConfig );
